refactor(tests): extract fixed_point_misc.cpp from fixed_point.cpp (part of #3130, meta #3127)
First extraction in the fixed_point.cpp split (#3130). Moves the misc
section (lines 4818-5226) into its own file:
  - Fixed-point scalar type alignment tests
  - Constexpr scalar multiply/divide tests (s0x32, u0x32)
  - Integer constructor tests (PR #2197 int width portability)
  - pow boundary regression tests (#2969)

Also extends ci/hooks/check_test_file_pairing.py to allow sibling-split
test files. The original 1:1 pairing rule rejected fixed_point_misc.cpp
because there is no src/fl/math/fixed_point_misc.h. The new sibling-split
allowance walks back the base name via underscore separators and allows
the write when the longest base prefix has both an existing paired test
file and a matching src header. This unblocks the rest of #3127's splits
without weakening the convention for genuinely unpaired test files.

Tests pass: bash test fl_math_fixed_point --cpp (1/1, 17/17 build steps).

fixed_point.cpp: 5,230 -> 4,816 LOC
fixed_point_misc.cpp (new): 440 LOC

Part of #3130. Meta tracking issue: #3127.
